add gw apis, move relays into CE
